Created base module and separated common.js into specific utils files
Also merged the remaining actually common stuff inside common into tools
and removed the lib folder
