+++ src/index.js
@@ -2,5 +2,7 @@
 
 import 'babel-polyfill';
 
+export createBaseModule from './modules/base';
 export createQueryModule from './modules/query';
+
 export default from './JoyMap';
